+++ src/index.ts
@@ -15,7 +15,6 @@ export {
   CdkDiffOptions,
   addCdkPipelineWorkflows,
   addActionBuildWorkflow,
-  addActionSynthWorkflow,
   addActionDeployWorkflow,
   addActionDiffWorkflow,
   addPrMainWorkflow,